Fix: Relocating Settings Screen
